+++ aiko/src/modules/display_module.h
@@ -16,7 +16,7 @@ namespace aiko
     
         bool isOpen();
 
-        vec2 getDisplaySize();
+        ivec2 getDisplaySize();
 
         void setWindowTitle(const char* title);                     // Set title for window (only PLATFORM_DESKTOP and PLATFORM_WEB)
         void setWindowPosition(int x, int y);                       // Set window position on screen (only PLATFORM_DESKTOP)
