+++ aiko/src/modules/renderer/renderer_component.h
@@ -32,8 +32,8 @@ namespace aiko
     
         virtual void dispose() { };
     
-        vec2 getDisplaySize();
-        virtual aiko::vec2 getDisplayViewport() { return { 0.0f, 0.0f }; };
+        ivec2 getDisplaySize();
+        virtual aiko::ivec2 getDisplayViewport() { return { 0, 0 }; };
     
     protected:
